adding token authentication and pagination

+++ mysite/blog/api/views.py
@@ -1,18 +1,20 @@
 from rest_framework import status
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import IsAuthenticated
 
+
+from rest_framework.pagination import PageNumberPagination
+from rest_framework.generics import ListAPIView
+from rest_framework.authentication import TokenAuthentication
 
 from account.models import Account
 from blog.models import BlogPost
 from blog.api.serializers import BlogPostSerializer
 
 
-
-
-
-
 @api_view(['GET',])
+@permission_classes((IsAuthenticated,))
 def api_detail_blog_view(request, slug):
 
     try:
@@ -20,18 +22,27 @@ def api_detail_blog_view(request, slug):
     except BlogPost.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
 
+
+
     if request.method == 'GET':
         serializer = BlogPostSerializer(blog_post)
         return Response(serializer.data)
 
 
 @api_view(['PUT',])
+@permission_classes((IsAuthenticated,))
 def api_update_blog_view(request, slug):
 
     try:
         blog_post = BlogPost.objects.get(slug=slug)
     except BlogPost.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
+
+
+    user = request.user
+    if blog_post.author != user:
+        return Response({'response' : 'You dont have permission to edit'})
+
 
     if request.method == 'PUT':
         serializer = BlogPostSerializer(blog_post, data=request.data)
@@ -43,12 +54,19 @@ def api_update_blog_view(request, slug):
         return Response(serializer.errors, status=status.HTTP_404_BAD_REQUEST)
 
 @api_view(['DELETE',])
+@permission_classes((IsAuthenticated,))
+
 def api_delete_blog_view(request, slug):
 
     try:
         blog_post = BlogPost.objects.get(slug=slug)
     except BlogPost.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
+
+    user = request.user
+    if blog_post.author != user:
+        return Response({'response' : 'You dont have permission to delete'})
+
 
     if request.method == 'DELETE':
         operation = blog_post.delete()
@@ -61,10 +79,12 @@ def api_delete_blog_view(request, slug):
         return Response(data=data)
 
 @api_view(['POST',])
+@permission_classes((IsAuthenticated,))
+
 def api_create_blog_view(request):
 
-    account = Account.objects.get(pk=1)
-
+    # account = Account.objects.get(pk=1)
+    account = request.user
     blog_post = BlogPost(author=account)
 
     if request.method == 'POST':
@@ -73,3 +93,12 @@ def api_create_blog_view(request):
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+
+class ApiBlogListView(ListAPIView):
+    queryset = BlogPost.objects.all()
+    serializer_class = BlogPostSerializer
+    authentication_classes = (TokenAuthentication,)
+    permission_classes = (IsAuthenticated,)
+    pagination_class = PageNumberPagination
